requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0

# AI Provider Libraries
anthropic>=0.18.1
google-generativeai>=0.4.0
openai>=1.12.0

# Web Framework and API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
Passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
PyJWT>=2.8.0

# Additional utilities
aiohttp>=3.9.3
pydantic>=2.6.1
pydantic-settings>=2.0.0
cryptography>=42.0.0
PyYAML>=6.0.1
httpx>=0.26.0
requests>=2.31.0

# System and monitoring
psutil>=5.9.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
